SWITCHYARD-2145: Security context not propagated to EJB invocations
