--- demos/policy-security-basic-propagate/src/main/java/org/switchyard/quickstarts/demo/policy/security/basic/propagate/WorkAck.java
@@ -1,48 +1,0 @@
-/*
- * JBoss, Home of Professional Open Source
- * Copyright 2013, Red Hat, Inc. and/or its affiliates, and individual
- * contributors by the @authors tag. See the copyright.txt in the
- * distribution for a full listing of individual contributors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.switchyard.quickstarts.demo.policy.security.basic.propagate;
-
-import java.io.Serializable;
-
-/**
- * @author David Ward &lt;<a href="mailto:dward@jboss.org">dward@jboss.org</a>&gt; &copy; 2013 Red Hat Inc.
- */
-@SuppressWarnings("serial")
-public class WorkAck implements Serializable {
-
-    private String _command;
-    private boolean _received;
-
-    public String getCommand() {
-        return _command;
-    }
-
-    public WorkAck setCommand(String command) {
-        _command = command;
-        return this;
-    }
-
-    public boolean isReceived() {
-        return _received;
-    }
-
-    public WorkAck setReceived(boolean received) {
-        _received = received;
-        return this;
-    }
-
-}
