--- demos/policy-security-basic-propagate/src/test/resources/log4j.xml
@@ -1,42 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-    JBoss, Home of Professional Open Source
-    Copyright 2013, Red Hat, Inc. and/or its affiliates, and individual
-    contributors by the @authors tag. See the copyright.txt in the
-    distribution for a full listing of individual contributors.
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-    http://www.apache.org/licenses/LICENSE-2.0
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
--->
-<!DOCTYPE log4j:configuration SYSTEM "log4j.dtd">
-<log4j:configuration xmlns:log4j='http://jakarta.apache.org/log4j/'>
-
-    <appender name="STDOUT" class="org.apache.log4j.ConsoleAppender">
-        <layout class="org.apache.log4j.PatternLayout">
-            <param name="ConversionPattern" value="%d{HH:mm:ss,SSS} %-5p [%c] %m%n"/>
-        </layout>
-    </appender>
-
-    <logger name="org.jboss.weld.Bootstrap">
-        <level value="ERROR"/>
-    </logger>
-    <logger name="org.apache.commons.httpclient.HttpMethodBase">
-        <level value="ERROR"/>
-    </logger>
-    <logger name="org.jboss.interceptor.util.InterceptionTypeRegistry">
-        <level value="ERROR"/>
-    </logger>
-
-    <root>
-        <level value="INFO"/>
-        <appender-ref ref="STDOUT"/>
-    </root>
-
-</log4j:configuration>
