--- demos/policy-security-basic-propagate/src/test/resources/xml/soap-request.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<soap:Envelope
-    xmlns:soap="http://schemas.xmlsoap.org/soap/envelope/"
-    xmlns:policy-security-basic-propagate="urn:switchyard-quickstart-demo:policy-security-basic-propagate:0.1.0">
-    <soap:Header/>
-    <soap:Body>
-        <policy-security-basic-propagate:doWork>
-            <work>
-                <command>WORK_CMD</command>
-            </work>
-        </policy-security-basic-propagate:doWork>
-    </soap:Body>
-</soap:Envelope>
