+++ demos/security-propagation/basic/src/main/resources/META-INF/switchyard.xml
@@ -1,0 +1,58 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+    JBoss, Home of Professional Open Source
+    Copyright 2014, Red Hat, Inc. and/or its affiliates, and individual
+    contributors by the @authors tag. See the copyright.txt in the
+    distribution for a full listing of individual contributors.
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+    http://www.apache.org/licenses/LICENSE-2.0
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+-->
+<switchyard xmlns="urn:switchyard-config:switchyard:1.0" name="securty-propagation-basic">
+    <composite xmlns="http://docs.oasis-open.org/ns/opencsa/sca/200912" name="security-propagation-basic" targetNamespace="urn:switchyard-quickstart-demo:security-propagation-basic:0.1.0">
+        <service name="WorkService" promote="WorkService">
+            <interface.wsdl interface="META-INF/WorkService.wsdl#wsdl.porttype(WorkService)"/>
+            <binding.soap xmlns="urn:switchyard-component-soap:config:1.0">
+                <wsdl>META-INF/WorkService.wsdl</wsdl>
+                <contextPath>security-propagation-basic</contextPath>
+            </binding.soap>
+        </service>
+        <component name="WorkService">
+            <implementation.bean xmlns="urn:switchyard-component-bean:config:1.0" class="org.switchyard.quickstarts.demo.security.propagation.basic.WorkServiceBean" requires="authorization"/>
+            <service name="WorkService" requires="confidentiality clientAuthentication">
+                <interface.java interface="org.switchyard.quickstarts.demo.security.propagation.basic.WorkService"/>
+            </service>
+            <reference name="BackEndService">
+                <interface.java interface="org.switchyard.quickstarts.demo.security.propagation.basic.BackEndService"/>
+            </reference>
+        </component>
+        <component name="BackEndService">
+            <implementation.bean xmlns="urn:switchyard-component-bean:config:1.0" class="org.switchyard.quickstarts.demo.security.propagation.basic.BackEndServiceBean" requires="authorization"/>
+            <service name="BackEndService" requires="confidentiality clientAuthentication">
+                <interface.java interface="org.switchyard.quickstarts.demo.security.propagation.basic.BackEndService"/>
+            </service>
+        </component>
+    </composite>
+    <transforms>
+        <transform.java xmlns="urn:switchyard-config:transform:1.0" class="org.switchyard.quickstarts.demo.security.propagation.basic.Transformers" from="java:org.switchyard.quickstarts.demo.security.propagation.basic.WorkAck" to="java:org.switchyard.quickstarts.demo.security.propagation.basic.Work"/>
+        <transform.java xmlns="urn:switchyard-config:transform:1.0" class="org.switchyard.quickstarts.demo.security.propagation.basic.Transformers" from="java:org.switchyard.quickstarts.demo.security.propagation.basic.WorkAck" to="{urn:switchyard-quickstart-demo:security-propagation-basic:0.1.0}doWorkResponse"/>
+        <transform.java xmlns="urn:switchyard-config:transform:1.0" class="org.switchyard.quickstarts.demo.security.propagation.basic.Transformers" from="{urn:switchyard-quickstart-demo:security-propagation-basic:0.1.0}doWork" to="java:org.switchyard.quickstarts.demo.security.propagation.basic.Work"/>
+    </transforms>
+    <domain>
+        <!--
+        <properties>
+            <property name="org.switchyard.handlers.messageTrace.enabled" value="true"/>
+        </properties>
+        -->
+        <securities>
+            <security rolesAllowed="friend"/>
+        </securities>
+    </domain>
+</switchyard>
