+++ demos/security-propagation/ejb/src/main/resources/META-INF/jboss-ejb3.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="UTF-8"?>  
+<jboss:ejb-jar xmlns:jboss="http://www.jboss.com/xml/ns/javaee"  
+    xmlns="http://java.sun.com/xml/ns/javaee" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"  
+    xmlns:security="urn:security" version="3.1" impl-version="2.0">  
+    <assembly-descriptor xmlns="http://java.sun.com/xml/ns/javaee">  
+        <!--
+        <security:security xmlns:security="urn:security">
+            <security:security-domain>other</security:security-domain>
+            <ejb-name>TestEJBBean</ejb-name>
+        </security:security>
+        -->
+    </assembly-descriptor>  
+</jboss:ejb-jar>
